feat: :sparkles: Implemented a Bulls & Cows (Wordle) Styled Answer Checker
Upon playtesting with my mom, I realized that there was no way for the player to know if they're on the right track or not, and even more so if they had one answer correct, prior the game would just state you're wrong, terrible user experience, just awful! I know I'd be upset. So as a solution I thought back to games such as Bulls and Cows, and Wordle, specifically Bulls and Cows, being a game project my group and I did for an assignment for one of our Web courses :D So now this update implements that, so the gameplay should be more forgiving and less frustrating